+++ R/register_all_transpilers.R
@@ -4,6 +4,7 @@ register_all_transpilers <- function() {
   ## Map-reduce packages (base-R)
   append_builtin_transpilers_for_base()
   append_builtin_transpilers_for_BiocParallel()
+  append_builtin_transpilers_for_crossmap()
   append_builtin_transpilers_for_future.apply()
   append_builtin_transpilers_for_purrr()
   append_builtin_transpilers_for_furrr()